fix(federation): remove duplicate types in schema (#1001)
This fix goes back to using  as used before 7.4.4 but
because of the bug present in the official Apollo package where
federation properties aren't carried over, this PR has that entire file
copied over locally with a somewhat hacky fix. Unless Apollo suddenly
changes federation however, the fix should be stable. The best solution
is of course that Apollo patches their package and we can revert to
using that, but I think for now this change is good to have, just so
all the problems that have been popping up (such as #1001 and #981) can
be fixed.