feat(plugin) process only classes with whitelisted decorators
Prior to this change typescript transform plugin processed all classes in
files with a matching names. This is probably not desirable behaviour
because there might be classes not exposed to graphql and they should not be processed

This change makes processing a bit smarter, so it only processes classes
decorated with ObjectType / InterfaceType and InputType decorators.